refactor map tooltip and label logic
